fix: ensure HADOLINT_PATH is available where used
Newer versions of `shellcheck` notices when aliases
are used in the same parsing context, exposing the
alias hack used to invoke custom hadolint paths.

The proper solution to allowing custom paths is simpler than that;
just read it from environment.

As part of this, switch from `:=` notation to `:-` for "set if empty".