fix: don't error if invocated twice
When running the action twice in a job, it fails because
the directory we tried to create already exists.

Check if there already is a binary and that it is the version
we expect. This also saves us a bit of time by avoiding a re-download.

Finally, add e2e tests to cover all use-cases.

PR: https://github.com/jbergstroem/hadolint-gh-action/pull/135
Closes: https://github.com/jbergstroem/hadolint-gh-action/issues/134
Co-Authored-By: Johan Bergström <bugs@bergstroem.nu>